+++ packages/type/tests/unit/ClassStringTypeTest.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace Psl\Type\Tests\Unit;
 
+use Exception;
 use Override;
+use PHPUnit\Framework\Attributes\DataProvider;
 use Psl\Collection;
 use Psl\Type;
+use Throwable;
 
 final class ClassStringTypeTest extends TypeTestCase
 {
@@ -44,5 +47,37 @@ final class ClassStringTypeTest extends TypeTestCase
         yield [Type\class_string(Collection\VectorInterface::class), 'class-string<Psl\Collection\VectorInterface>'];
         yield [Type\class_string(Collection\Vector::class), 'class-string<Psl\Collection\Vector>'];
         yield [Type\class_string(Collection\Map::class), 'class-string<Psl\Collection\Map>'];
+        yield [Type\class_string(), 'class-string'];
+    }
+
+    public static function validValuesForUnrestrictedType(): iterable
+    {
+        yield [Collection\Vector::class];
+        yield [Collection\MutableVector::class];
+        yield [Collection\Map::class];
+        yield [Collection\MutableMap::class];
+        yield [Exception::class];
+    }
+
+    #[DataProvider('validValuesForUnrestrictedType')]
+    public function testUnspecifiedTypeAcceptsAnyClassString(string $value): void
+    {
+        static::assertSame($value, Type\class_string()->assert($value));
+    }
+
+    /** @return iterable<array{0: mixed}> */
+    public static function invalidValuesForUnspecifiedType(): iterable
+    {
+        yield from self::getInvalidCoercions();
+        yield [Collection\MutableMapInterface::class];
+        yield [Collection\CollectionInterface::class];
+        yield [Throwable::class];
+    }
+
+    #[DataProvider('invalidValuesForUnspecifiedType')]
+    public function testInvalidValuesWhenTypeIsUnspecified(mixed $value): void
+    {
+        $this->expectException(Type\Exception\AssertException::class);
+        Type\class_string()->assert($value);
     }
 }